pyproject.toml
[tool.poetry]
name = "egon"
version = "0.1.0"
description = "EnerSHelf"
authors = ["Hendrik Huyskens <hendrik.huyskens@rl-institut.de>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
gunicorn = "^20.0.4"  # https://github.com/benoitc/gunicorn
# pytz = "^2021.1"  # https://github.com/stub42/pytz
# python-slugify = "^5.0.2"  # https://github.com/un33k/python-slugify
# Pillow = "^8.2.0"  # https://github.com/python-pillow/Pillow
rcssmin = "^1.0.6"  # https://github.com/ndparker/rcssmin
argon2-cffi = "^20.1.0"  # https://github.com/hynek/argon2_cffi
whitenoise = "^5.2.0"  # https://github.com/evansd/whitenoise
redis = "^3.5.3"  # https://github.com/antirez/redis
range-key-dict = "^1.1.0"
psycopg2 = "^2.8.6" # https://github.com/psycopg/psycopg2

# Django
# ------------------------------------------------------------------------------
django = "^3.2.3"  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ = "^0.10.0"  # https://github.com/joke2k/django-environ
# django-model-utils = "^4.1.1"  # https://github.com/jazzband/django-model-utils
django-allauth = "^0.44.0"  # https://github.com/pennersr/django-allauth
django-foundation-formtags = "^0.1.0"
django-compressor = "^2.4.1"  # https://github.com/django-compressor/django-compressor
django-libsass = "^0.8"
django-redis = "^4.12.1"  # https://github.com/niwinz/django-redis
django-distill = "^2.6"
django-crispy-forms = "^1.11.2"
django-geojson = "^3.1.0"
django-mapengine = {git = "https://github.com/rl-institut/django-mapengine.git"}

# Django REST Framework
djangorestframework = "^3.12.4"  # https://github.com/encode/django-rest-framework
coreapi = "^2.3.3"  # https://github.com/core-api/python-client
django-select2 = "^7.7.1"
geojson = "^2.5.0"

[tool.poetry.dev-dependencies]
Werkzeug = "^2.0.1" # https://github.com/pallets/werkzeug
ipdb = "^0.13.7"  # https://github.com/gotcha/ipdb
Sphinx = "^4.0.2"  # https://github.com/sphinx-doc/sphinx

# Testing
# ------------------------------------------------------------------------------
# mypy = "^0.812"  # https://github.com/python/mypy
pytest = "^6.2.4"  # https://github.com/pytest-dev/pytest
pytest-sugar = "^0.9.4"  # https://github.com/Frozenball/pytest-sugar

# Code quality
# ------------------------------------------------------------------------------
flake8 = "^3.9.2"  # https://github.com/PyCQA/flake8
black = "^22.3.0"  # https://github.com/ambv/black
pylint-django = "^2.4.4"  # https://github.com/PyCQA/pylint-django
pre-commit = "^2.13.0"  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
# factory-boy = 3.2.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar = "^3.2.1"  # https://github.com/jazzband/django-debug-toolbar
django-extensions = "^3.1.3"  # https://github.com/django-extensions/django-extensions
# django-coverage-plugin = "^1.6.0"  # https://github.com/nedbat/django_coverage_plugin
pytest-django = "^4.3.0"  # https://github.com/pytest-dev/pytest-django

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
